+++ resources/views/platforms/index.blade.php
@@ -2,18 +2,20 @@
 
 @section('content')
     <div class="top_title">
-       @include('master.breadcrumb', [
+        @include('master.breadcrumb', [
             'title' => 'Platforms',
             'icon' => '',
             'sub_title' => [
                 'Enox ERP Platforms' => '',
             ],
         ])
-        <div>
-            <a href="{{ route('admin.platforms.create') }}" class="btn btn-outline-secondary">
-                Create <span><i class="bi bi-plus-lg me-0"></i></span>
-            </a>
-        </div>
+        @can('settings.platforms.create')
+            <div>
+                <a href="{{ route('admin.platforms.create') }}" class="btn btn-outline-secondary">
+                    Create <span><i class="bi bi-plus-lg me-0"></i></span>
+                </a>
+            </div>
+        @endcan
     </div>
     <form method="GET" action="{{ route('admin.platforms.index') }}">
         <div class="card" id="filterSection">
@@ -26,13 +28,15 @@
                     </div>
                     <div class="col-12 col-md-4">
                         <div class="form-group mb-3 mb-md-0 new_select_field new_same_item d-flex flex-wrap">
-                            <input type="text" name="q" id="q" class="form-control" placeholder="Search here...." value="{{request('q')}}" />
+                            <input type="text" name="q" id="q" class="form-control"
+                                placeholder="Search here...." value="{{ request('q') }}" />
                         </div>
                     </div>
                     <div class="col-12 col-md-8 text-end mt-2 mt-md-0">
                         <div class="flex-center">
                             <a href="{{ route('admin.platforms.index') }}"
-                                class="btn btn-outline-danger flex-center mx-1"><i class="bi bi-arrow-clockwise ms-0"></i> Reset</a>
+                                class="btn btn-outline-danger flex-center mx-1"><i class="bi bi-arrow-clockwise ms-0"></i>
+                                Reset</a>
                             <button type="submit" class="btn btn-primary mx-1"><i class="fa fa-filter ms-0"
                                     aria-hidden="true"></i>
                                 Search</button>
@@ -64,28 +68,31 @@
                                     <td>{{ $data->name }}</td>
                                     <td>@price($data->shipping_charge)</td>
                                     <td>
-                                        CA: {{ $data->created_at }} <br/>
+                                        CA: {{ $data->created_at }} <br />
                                         UA: {{ $data->updated_at }}
                                     </td>
                                     <td>{{ $data->note }}</td>
                                     <td class="text-center">
                                         <div class="d-inline-flex text-center text-md-start text-nowrap">
-                                            <a class="btn btn-soft-primary btn-sm mx-1" title="Edit"
-                                                href="{{ route('admin.platforms.edit', $data->id) }}">
-                                                <iconify-icon icon="solar:pen-2-broken" class="fs-18"></iconify-icon>
-                                            </a>
-
-                                            <button class="btn btn-soft-danger btn-sm delete-btn mx-1" type="button"
-                                                onclick="deleteData({{ $data->id }})" data-id="{{ $data->id }}">
-                                                <iconify-icon icon="solar:trash-bin-minimalistic-2-broken"
-                                                    class="fs-18 delete-icon"></iconify-icon>
-                                            </button>
-                                            <form id="delete-form-{{ $data->id }}" method="POST"
-                                                action="{{ route('admin.platforms.destroy', $data->id) }}"
-                                                style="display: none;">
-                                                @csrf
-                                                @method('DELETE')
-                                            </form>
+                                            @can('settings.platforms.edit')
+                                                <a class="btn btn-soft-primary btn-sm mx-1" title="Edit"
+                                                    href="{{ route('admin.platforms.edit', $data->id) }}">
+                                                    <iconify-icon icon="solar:pen-2-broken" class="fs-18"></iconify-icon>
+                                                </a>
+                                            @endcan
+                                            @can('settings.platforms.delete')
+                                                <button class="btn btn-soft-danger btn-sm delete-btn mx-1" type="button"
+                                                    onclick="deleteData({{ $data->id }})" data-id="{{ $data->id }}">
+                                                    <iconify-icon icon="solar:trash-bin-minimalistic-2-broken"
+                                                        class="fs-18 delete-icon"></iconify-icon>
+                                                </button>
+                                                <form id="delete-form-{{ $data->id }}" method="POST"
+                                                    action="{{ route('admin.platforms.destroy', $data->id) }}"
+                                                    style="display: none;">
+                                                    @csrf
+                                                    @method('DELETE')
+                                                </form>
+                                            @endcan
                                         </div>
                                     </td>
                                 </tr>
